commands: move restore code to restore.rs
